Fix AMP dtype mismatches at module boundaries
Under use_amp=True, model features/logits arrive as float16; convs called
outside an autocast context crashed (Input type Half vs weight float). Fixes:
- DomainDiscriminator / PatchDiscriminator cast input to float32
- entropy_map and CBST pseudo-labeling compute in float32
- RenyiLoss computes internally in float32 and DARES engine runs the
  alignment outside autocast (Gram-matrix trace normalizations are precision
  sensitive)
- Add tests/test_amp_safety.py regressing the half-input boundaries

+++ src/dares/losses/advent.py
@@ -29,7 +29,7 @@ def entropy_map(logits: torch.Tensor) -> torch.Tensor:
         (nats), where ``1`` denotes maximal uncertainty and ``0`` perfect
         confidence.
     """
-    p = F.softmax(logits, dim=1)
+    p = F.softmax(logits.float(), dim=1)
     ent = -(p * (p + _EPS).log()).sum(dim=1, keepdim=True)
     return ent
